1.1.0
Fixed naming on existing weapons.
Fixed the caliber of the ejected casings for the AF.
Fixed the caliber of the ejected casings for the AP67.
Fixed the caliber of the ejected casings for the .308 Tigr.
Adjusted shot sounds for the AF.
Adjusted shot sounds for the AP67.
Added Ruger Mini-14.
Added Ruger AC-556.
Added AKS-74.
Added AKS-74N.
Added the ability to launch a MPL-50 Entrenching Tool out of the RPG-7 (I got bored).